Port over git utilities
