+++ Cargo.toml
@@ -14,6 +14,7 @@ uv-client = { git = "https://github.com/astral-sh/uv", tag = "0.5.29" }
 uv-configuration = { git = "https://github.com/astral-sh/uv", tag = "0.5.29" }
 uv-dispatch = { git = "https://github.com/astral-sh/uv", tag = "0.5.29" }
 uv-distribution = { git = "https://github.com/astral-sh/uv", tag = "0.5.29" }
+uv-distribution-filename = { git = "https://github.com/astral-sh/uv", tag = "0.5.29" }
 uv-distribution-types = { git = "https://github.com/astral-sh/uv", tag = "0.5.29" }
 uv-install-wheel = { git = "https://github.com/astral-sh/uv", tag = "0.5.29" }
 uv-installer = { git = "https://github.com/astral-sh/uv", tag = "0.5.29" }
@@ -43,6 +44,7 @@ reqwest = "0.12.12"
 tokio = { version = "1.43.0", features = ["fs"] }
 tracing = "0.1.41"
 url = "2.5.4"
+percent-encoding = "2.3.1"
 
 [dependencies.pyo3]
 version = "0.22.6"
